+++ app/src/main/java/br/com/angelorobson/templatemvi/model/dtos/SpotlightDto.kt
@@ -8,7 +8,7 @@ data class SpotlightDto(
         val title: String,
         val publisher: String,
         val image: String,
-        val discount: Int,
+        val discount: Double,
         val price: Double,
         val description: String,
         val rating: Double,
